+++ tests/string.es6.test.js
@@ -183,3 +183,15 @@ describe("String.prototype.words", () => {
       .toEqual(['python', 'javaScript','coffee'])
   })
 })
+
+describe("String.prototype.contains", () => {
+  test("'javaScript & typescript'.contains('Typescript'); // true", () => {
+    expect('javaScript & typescript'.contains('Typescript'))
+      .toBe(true)
+  })
+
+  test("'javaScript & typescript'.contains('Typescript', true); // false", () => {
+    expect('javaScript & typescript'.contains('Typescript', true))
+      .toBe(false)
+  })
+})